AES/DAGpt1/airflow-docker/dags/requirements.txt
alembic==1.8.0
anyio==3.6.1
apache-airflow-backport-providers-apache-spark==2021.3.3
apache-airflow-providers-ftp==3.0.0
apache-airflow-providers-http==3.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.0.0
apache-airflow==1.10.15
apispec==1.3.3
argcomplete==1.12.3
attrs==20.3.0
Babel==2.10.3
blinker==1.4
cached-property==1.5.2
cachelib==0.8.0
cattrs==1.10.0
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
colorama==0.4.5
colorlog==4.0.2
commonmark==0.9.1
configparser==3.5.3
connexion==2.14.0
cron-descriptor==1.2.24
croniter==0.3.37
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==0.14.3
Flask==1.1.4
funcsigs==1.0.2
future==0.18.2
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==2.10
importlib-metadata==2.1.3
importlib-resources==1.5.0
inflection==0.5.1
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.2.0
Markdown==2.6.11
MarkupSafe==2.0.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
marshmallow==2.21.0
natsort==8.1.0
numpy==1.23.0
package-name==0.1
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pendulum==1.4.4
pip==22.1.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.1
py4j==0.10.9.5
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
requests==2.23.0
rfc3986==1.5.0
rich==12.4.4
setproctitle==1.2.3
setuptools==49.2.1
six==1.15.0
sniffio==1.2.0
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.38.2
SQLAlchemy==1.4.9
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.3
thrift==0.16.0
typing-extensions==4.2.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.11
Werkzeug==0.16.1
wheel==0.36.2
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
zope.deprecation==4.4.0
